requirements.txt
base58==2.1.1
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
ipfsapi==0.4.4
ipfshttpclient==0.4.13.2
multiaddr==0.0.9
netaddr==0.8.0
pymilter==1.0.5
requests==2.28.1
six==1.16.0
urllib3==1.26.13
varint==1.0.2
